talktv/requirements.txt
Django>=1.11,<1.12
djangorestframework==3.7.1
MySQL-python==1.2.5
python-decouple
django-unixdatetimefield==0.1.6
python-dateutil
pymongo
django-cors-headers
google-api-python-client
python-ldap
gevent
xlsxwriter
pandas